refactor(#20): split colorconverter in multiple components
